Move ballot-interpreter-vx code to ballot-interpreter
