+++ apps/central-scan/frontend/tsconfig.json
@@ -19,7 +19,6 @@
   "references": [
     { "path": "../../../libs/api/tsconfig.build.json" },
     { "path": "../../../libs/ballot-encoder/tsconfig.build.json" },
-    { "path": "../../../libs/ballot-interpreter-vx/tsconfig.build.json" },
     { "path": "../../../libs/basics/tsconfig.build.json" },
     { "path": "../../../libs/dev-dock/frontend/tsconfig.build.json" },
     { "path": "../../../libs/eslint-plugin-vx/tsconfig.build.json" },
