+++ apps/scan/backend/package.json
@@ -41,7 +41,6 @@
     "@votingworks/backend": "workspace:*",
     "@votingworks/ballot-encoder": "workspace:*",
     "@votingworks/ballot-interpreter": "workspace:*",
-    "@votingworks/ballot-interpreter-vx": "workspace:*",
     "@votingworks/basics": "workspace:*",
     "@votingworks/converter-nh-accuvote": "workspace:*",
     "@votingworks/custom-scanner": "workspace:*",
