+++ apps/scan/backend/tsconfig.json
@@ -21,7 +21,6 @@
     { "path": "../../../libs/auth/tsconfig.build.json" },
     { "path": "../../../libs/ballot-encoder/tsconfig.build.json" },
     { "path": "../../../libs/ballot-interpreter/tsconfig.build.json" },
-    { "path": "../../../libs/ballot-interpreter-vx/tsconfig.build.json" },
     { "path": "../../../libs/basics/tsconfig.build.json" },
     { "path": "../../../libs/backend/tsconfig.build.json" },
     { "path": "../../../libs/converter-nh-accuvote/tsconfig.build.json" },
