--- libs/ballot-interpreter-vx/.eslintignore
@@ -1,4 +1,0 @@
-bin
-build
-coverage
-jest.config.js
