--- libs/ballot-interpreter-vx/.eslintrc
@@ -1,14 +1,0 @@
-{
-  "extends": ["plugin:vx/recommended"],
-  "rules": {
-    // Allow specific exceptions to casing rules.
-    "camelcase": ["error", { "allow": ["matrix_t", "^.+_\\d+"] }],
-    "new-cap": ["error", { "newIsCapExceptionPattern": "^jsfeat\\.." }],
-    // Allow for bitwise operation which is necessary for encoding ballots.
-    "no-bitwise": "off",
-    // $0 is often used as the name of the current process by convention.
-    "vx/gts-identifiers": ["error", { "allowedNames": ["$0"] }],
-    // Disable JSDOC rule as code is self-documenting.
-    "vx/gts-jsdoc": "off"
-  }
-}
