--- libs/ballot-interpreter-vx/.gitignore
@@ -1,12 +1,0 @@
-coverage
-node_modules
-*.d.ts
-*.d.ts.map
-*.js
-!*.config.js
-!*rc.js
-!types/**/*.d.ts
-!bin/*.js
-
-debug*.png
-votingworks-ballot-interpreter-vx*.tgz
