--- libs/ballot-interpreter-vx/README.md
@@ -1,11 +1,0 @@
-# ballot-interpreter-vx
-
-Interprets VotingWorks machine-marked ballots.
-
-## Setup
-
-Follow the instructions in the [VxSuite README](../../README.md) to get set up.
-
-## License
-
-AGPL-3.0
