--- libs/ballot-interpreter-vx/bin/ballot-interpreter-vx.js
@@ -1,13 +1,0 @@
-#!/usr/bin/env node
-
-require('esbuild-runner/register');
-
-require('../src/cli')
-  .main(process.argv, process.stdin, process.stdout, process.stderr)
-  .then((code) => {
-    process.exitCode = code;
-  })
-  .catch((error) => {
-    console.error(error);
-    process.exit(1);
-  });
