--- libs/ballot-interpreter-vx/jest.config.js
@@ -1,16 +1,0 @@
-const shared = require('../../jest.config.shared');
-
-/**
- * @type {import('@jest/types').Config.InitialOptions}
- */
-module.exports = {
-  ...shared,
-  coverageThreshold: {
-    global: {
-      statements: 90,
-      branches: 60,
-      functions: 80,
-      lines: 90,
-    },
-  },
-};
