--- libs/ballot-interpreter-vx/package.json
@@ -1,81 +1,0 @@
-{
-  "name": "@votingworks/ballot-interpreter-vx",
-  "version": "5.2.16",
-  "description": "Interprets hand-marked paper ballots.",
-  "repository": "https://github.com/votingworks/ballot-interpreter-vx",
-  "license": "GPL-3.0",
-  "author": "VotingWorks Eng <eng@voting.works>",
-  "main": "build/index.js",
-  "types": "build/index.d.ts",
-  "bin": {
-    "ballot-interpreter-vx": "bin/ballot-interpreter-vx.js"
-  },
-  "files": [
-    "build"
-  ],
-  "scripts": {
-    "type-check": "tsc --build",
-    "build": "tsc --build tsconfig.build.json",
-    "clean": "rm -rf build tsconfig.tsbuildinfo tsconfig.build.tsbuildinfo",
-    "lint": "pnpm type-check && eslint .",
-    "lint:fix": "pnpm type-check && eslint . --fix",
-    "prepack": "tsc --build",
-    "test": "is-ci test:ci test:watch",
-    "test:ci": "pnpm test:coverage --reporters=default --reporters=jest-junit --maxWorkers=6",
-    "test:coverage": "jest --coverage",
-    "test:watch": "jest --watch",
-    "pre-commit": "lint-staged"
-  },
-  "lint-staged": {
-    "*.+(css|graphql|json|less|md|mdx|sass|scss|yaml|yml)": [
-      "prettier --write"
-    ],
-    "*.+(js|jsx|ts|tsx)": [
-      "eslint --quiet --fix"
-    ],
-    "package.json": [
-      "sort-package-json"
-    ]
-  },
-  "dependencies": {
-    "@votingworks/ballot-encoder": "workspace:*",
-    "@votingworks/basics": "workspace:*",
-    "@votingworks/image-utils": "workspace:*",
-    "@votingworks/qrdetect": "^1.0.1",
-    "@votingworks/types": "workspace:*",
-    "@votingworks/utils": "workspace:*",
-    "buffer": "^6.0.3",
-    "canvas": "2.9.1",
-    "debug": "^4.2.0",
-    "jsfeat": "^0.0.8",
-    "jsqr": "^1.3.1",
-    "node-quirc": "^2.2.1",
-    "util": "^0.12.4",
-    "uuid": "^8.3.1"
-  },
-  "devDependencies": {
-    "@types/benchmark": "^1.0.33",
-    "@types/debug": "^4.1.5",
-    "@types/jest": "^29.5.3",
-    "@types/jsfeat": "workspace:*",
-    "@types/node": "16.18.23",
-    "@types/node-quirc": "workspace:*",
-    "@types/tmp": "^0.2.0",
-    "@types/uuid": "^8.3.0",
-    "@votingworks/fixtures": "workspace:*",
-    "benchmark": "^2.1.4",
-    "esbuild": "^0.14.25",
-    "esbuild-runner": "^2.2.1",
-    "eslint-plugin-vx": "workspace:*",
-    "is-ci-cli": "^2.2.0",
-    "jest": "^29.6.2",
-    "jest-junit": "^16.0.0",
-    "jest-watch-typeahead": "^2.2.2",
-    "js-sha256": "^0.9.0",
-    "lint-staged": "^10.4.2",
-    "sort-package-json": "^1.50.0",
-    "tmp": "^0.2.1",
-    "ts-jest": "^29.1.1"
-  },
-  "packageManager": "pnpm@8.1.0"
-}
