--- libs/ballot-interpreter-vx/tsconfig.build.json
@@ -1,20 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "include": ["src"],
-  "exclude": ["**/*.test.ts", "**/*.test.tsx"],
-  "compilerOptions": {
-    "noEmit": false,
-    "rootDir": "src",
-    "outDir": "build",
-    "declaration": true
-  },
-  "references": [
-    { "path": "../ballot-encoder/tsconfig.build.json" },
-    { "path": "../basics/tsconfig.build.json" },
-    { "path": "../eslint-plugin-vx/tsconfig.build.json" },
-    { "path": "../fixtures/tsconfig.build.json" },
-    { "path": "../image-utils/tsconfig.build.json" },
-    { "path": "../types/tsconfig.build.json" },
-    { "path": "../utils/tsconfig.build.json" }
-  ]
-}
