--- libs/ballot-interpreter-vx/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "extends": "../../tsconfig.shared.json",
-  "include": ["src", "test"],
-  "exclude": [],
-  "compilerOptions": {
-    "module": "commonjs",
-    "lib": ["ES2020"],
-    "composite": true,
-    "noEmit": true,
-    "noUnusedParameters": true,
-    "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true,
-    "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "noUncheckedIndexedAccess": false
-  },
-  "references": [
-    { "path": "../ballot-encoder/tsconfig.build.json" },
-    { "path": "../basics/tsconfig.build.json" },
-    { "path": "../eslint-plugin-vx/tsconfig.build.json" },
-    { "path": "../fixtures/tsconfig.build.json" },
-    { "path": "../image-utils/tsconfig.build.json" },
-    { "path": "../types/tsconfig.build.json" },
-    { "path": "../utils/tsconfig.build.json" }
-  ]
-}
