+++ libs/ballot-interpreter/README.md
@@ -1,6 +1,7 @@
 # @votingworks/ballot-interpreter
 
-Wrapper package for our various ballot interpreters.
+Library to interpret BMD ballots produced by VxMark or hand-marked paper ballots
+(HMPB) produced by VxDesign.
 
 ## License
