+++ libs/ballot-interpreter/src/bmd/index.ts
