+++ libs/ballot-interpreter/src/bmd/interpret.ts
