+++ libs/ballot-interpreter/src/bmd/types.ts
