+++ libs/ballot-interpreter/src/bmd/utils/qrcode.test.ts
